requirements.txt
torch==2.3.1
torchvision==0.18.1
pytorch-lightning==2.3.0
hydra-core==1.3.2
omegaconf==2.3.0
scikit-learn==1.5.0
matplotlib==3.9.0
tqdm==4.66.4
tensorboard==2.17.0
albumentations==1.2.1
opencv-python==4.5.3.56
wandb==0.17.3
